requirements.txt
inquirerPy
fastapi
psycopg2-binary
python-dotenv
requests
tabulate
